backend/utils/suspect_tracker.py
import cv2
import os
import numpy as np
from typing import Dict, List, Any, Optional
import asyncio
import logging
from PIL import Image
import time

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# In a real implementation, this would use a proper face/person detection and recognition model
# For now, we'll simulate the detection process

class SuspectTracker:
    def __init__(self):
        """
        Initialize the suspect tracker
        In a production system, this would load ML models for face/person detection and recognition
        """
        logger.info("Initializing SuspectTracker")
        # Simulate loading detection models
        self.detection_model_loaded = True
        self.recognition_model_loaded = True
        
        # In a real implementation, these would be actual ML models
        # self.detector = load_detection_model()
        # self.recognizer = load_recognition_model()
    
    async def detect_persons(self, frame: np.ndarray) -> List[Dict[str, Any]]:
        """
        Detect persons in a frame
        
        Args:
            frame: Image frame as numpy array
        
        Returns:
            List of detected persons with bounding boxes and confidence scores
        """
        # Simulate person detection
        # In a real implementation, this would use a proper detection model
        
        # For simulation, we'll randomly generate some detections
        height, width = frame.shape[:2]
        num_detections = np.random.randint(0, 3)  # 0-2 detections per frame
        
        detections = []
        for i in range(num_detections):
            # Generate random bounding box
            x1 = np.random.randint(0, width // 2)
            y1 = np.random.randint(0, height // 2)
            w = np.random.randint(width // 4, width // 2)
            h = np.random.randint(height // 4, height // 2)
            
            # Ensure box is within frame
            x2 = min(x1 + w, width)
            y2 = min(y1 + h, height)
            
            # Generate random confidence score
            confidence = np.random.uniform(0.6, 0.95)
            
            detections.append({
                "bbox": [x1, y1, x2, y2],
                "confidence": confidence
            })
        
        return detections
    
    async def compute_embedding(self, frame: np.ndarray, bbox: List[int]) -> np.ndarray:
        """
        Compute facial/person embedding for recognition
        
        Args:
            frame: Image frame as numpy array
            bbox: Bounding box [x1, y1, x2, y2]
        
        Returns:
            Embedding vector as numpy array
        """
        # Simulate computing embedding
        # In a real implementation, this would use a proper face/person recognition model
        
        # For simulation, we'll generate a random embedding vector
        embedding_size = 128
        embedding = np.random.normal(0, 1, embedding_size)
        
        # Normalize embedding
        embedding = embedding / np.linalg.norm(embedding)
        
        return embedding
    
    async def compare_embeddings(self, embedding1: np.ndarray, embedding2: np.ndarray) -> float:
        """
        Compare two embeddings and return similarity score
        
        Args:
            embedding1: First embedding vector
            embedding2: Second embedding vector
        
        Returns:
            Similarity score (0-1, higher is more similar)
        """
        # Compute cosine similarity
        similarity = np.dot(embedding1, embedding2)
        
        # Ensure similarity is between 0 and 1
        similarity = max(0, min(1, similarity))
        
        return similarity

# Global tracker instance
tracker = SuspectTracker()

async def track_suspect(
    suspect: Dict[str, Any],
    videos: List[Dict[str, Any]],
    timeframe: Optional[Dict[str, str]] = None
) -> List[Dict[str, Any]]:
    """
    Track a suspect across multiple videos
    
    Args:
        suspect: Suspect data including image URL
        videos: List of video data
        timeframe: Optional timeframe filter
    
    Returns:
        List of tracking results with timestamps and confidence scores
    """
    logger.info(f"Tracking suspect {suspect['id']} across {len(videos)} videos")
    
    # Load suspect image
    suspect_image_path = f"data/{suspect['imageUrl'].lstrip('/')}"
    
    if not os.path.exists(suspect_image_path):
        logger.error(f"Suspect image not found: {suspect_image_path}")
        return []
    
    # Load suspect image and convert to numpy array
    suspect_img = np.array(Image.open(suspect_image_path))
    
    # Compute suspect embedding
    # In a real implementation, we would detect the face/person first
    height, width = suspect_img.shape[:2]
    suspect_bbox = [0, 0, width, height]  # Use entire image
    suspect_embedding = await tracker.compute_embedding(suspect_img, suspect_bbox)
    
    # Track suspect across videos
    tracking_results = []
    
    for video in videos:
        # Get video frames directory
        frames_dir = f"data/videos/frames/{video['id']}"
        
        # Check if frames have been extracted
        if not os.path.exists(frames_dir):
            logger.warning(f"Frames not found for video {video['id']}, skipping")
            continue
        
        # Get list of frame files
        frame_files = sorted([f for f in os.listdir(frames_dir) if f.endswith('.jpg')])
        
        # Process each frame
        for i, frame_file in enumerate(frame_files):
            frame_path = os.path.join(frames_dir, frame_file)
            
            # Load frame
            frame = np.array(Image.open(frame_path))
            
            # Detect persons in frame
            detections = await tracker.detect_persons(frame)
            
            # For each detection, check if it matches the suspect
            for j, detection in enumerate(detections):
                # Compute embedding for the detected person
                person_embedding = await tracker.compute_embedding(frame, detection['bbox'])
                
                # Compare with suspect embedding
                similarity = await tracker.compare_embeddings(suspect_embedding, person_embedding)
                
                # If similarity is above threshold, consider it a match
                threshold = 0.7  # Adjust as needed
                if similarity >= threshold:
                    # Calculate timestamp based on frame index
                    # Assuming frames are extracted at 1 frame per second
                    seconds = i
                    video_timestamp = video['timestamp']
                    
                    # Convert video timestamp to datetime
                    from datetime import datetime, timedelta
                    base_time = datetime.fromisoformat(video_timestamp.replace('Z', '+00:00'))
                    event_time = base_time + timedelta(seconds=seconds)
                    event_timestamp = event_time.isoformat()
                    
                    # Create tracking result
                    result = {
                        "id": f"event-{video['id']}-{i}-{j}",
                        "suspectId": suspect['id'],
                        "videoId": video['id'],
                        "timestamp": event_timestamp,
                        "confidence": similarity * 100,  # Convert to percentage
                        "frame_index": i,
                        "bbox": detection['bbox'],
                        "startTime": seconds,
                        "endTime": seconds + 5,  # Assume 5-second clip
                    }
                    
                    tracking_results.append(result)
    
    # Sort results by timestamp
    tracking_results.sort(key=lambda x: x['timestamp'])
    
    # Apply timeframe filter if provided
    if timeframe:
        from datetime import datetime
        start_time = datetime.fromisoformat(timeframe['start'].replace('Z', '+00:00'))
        end_time = datetime.fromisoformat(timeframe['end'].replace('Z', '+00:00'))
        
        tracking_results = [
            result for result in tracking_results
            if start_time <= datetime.fromisoformat(result['timestamp'].replace('Z', '+00:00')) <= end_time
        ]
    
    logger.info(f"Found {len(tracking_results)} suspect appearances")
    
    return tracking_results
