backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
opencv-python==4.9.0.80
numpy==1.26.4
qdrant-client==1.7.3
pymongo==4.6.1
python-dotenv==1.0.1
pydantic==2.6.3
pillow==10.2.0
httpx==0.27.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
